# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/documents": {
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create document (optionally inline content, auto-finalizes)",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Restrict to one project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List documents in workspace",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/tree": {
      "get": {
        "operationId": "DocumentsController_tree",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Restrict the tree to one project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Document tree (feature 08 nesting); scoped to one project when projectId is given",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get document + its head (or given) revision + chunk summaries",
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Update title / category / parent (features 07 + 08); parentId: null re-roots",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/content": {
      "get": {
        "operationId": "DocumentsController_content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Full raw content of a revision (feature 01; head of the default branch by default)",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/graph": {
      "get": {
        "operationId": "DocumentsController_graph",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "depth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Document neighbourhood in the knowledge graph (feature 06); depth = entity hops (1-3)",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/revisions": {
      "get": {
        "operationId": "DocumentsController_listRevisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List revisions with DAG parentage (optionally filtered by branch)",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_createRevision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Optimistic concurrency (plan.md §7): expected branch-head revision id; 409 + comparison link when the head advanced",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRevisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create new draft revision on branch head (If-Match: expected head revision id)",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/branches": {
      "post": {
        "operationId": "DocumentsController_createBranch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create a branch from a revision (defaults to default-branch head)",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "DocumentsController_listBranches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List branches with head revisions",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/compare": {
      "get": {
        "operationId": "DocumentsController_compare",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "structural",
            "required": false,
            "in": "query",
            "description": "Include path-level structural diff (default true)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "semantic",
            "required": false,
            "in": "query",
            "description": "Include graph-projection semantic diff (default false)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "direct",
                "merge-base"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Compare two revisions (direct or merge-base mode, plan.md §8)",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/documents/{id}/relations/curate": {
      "post": {
        "operationId": "DocumentsController_curateRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurateRelationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Curate a relation: user-confirmed, confidence 1, protected from re-extraction (plan.md §5)",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/uploads": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create (or reuse) draft revision + get presigned PUT URL",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{id}/revisions/{revisionId}/finalize": {
      "post": {
        "operationId": "DocumentsController_finalize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Finalize an uploaded revision: hash, dedupe, enqueue indexing",
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Knowledge API",
    "description": "Dynamic Knowledge Platform — documents, revisions, ingestion, search",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "kn_ API key or ks_ session token (AUTH_MODE=api-key)"
      }
    },
    "schemas": {
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "description": "Owning project; must belong to workspaceId"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/InlineContentDto"
          },
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationInputDto"
            }
          },
          "category": {
            "type": "string",
            "enum": [
              "process",
              "use-case",
              "contract",
              "erd",
              "architecture",
              "guide",
              "reference",
              "other"
            ],
            "default": "other",
            "description": "Feature 07 categorization"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "description": "Feature 08 nesting: create under this document"
          }
        },
        "required": [
          "workspaceId",
          "projectId",
          "title"
        ]
      },
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "process",
              "use-case",
              "contract",
              "erd",
              "architecture",
              "guide",
              "reference",
              "other"
            ]
          },
          "parentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateRevisionDto": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string",
            "default": "main"
          },
          "message": {
            "type": "string"
          },
          "contentType": {
            "type": "string",
            "example": "text/markdown"
          }
        }
      },
...
    }
  }
}